pyproject.toml
[project]
name = "arc-agi-3-agents"
version = "0.1.0"
description = "ARC-AGI-3-Agents"
readme = "readme.md"
requires-python = ">=3.12"
dependencies = [
    "arc-agi>=0.9.1",
    "dotenv>=0.9.9",
    "flask",
    "langchain",
    "langchain-core",
    "langchain-google-genai",
    "langchain-openai",
    "langgraph",
    "langgraph-checkpoint-sqlite",
    "numpy>=2.3.2",
    "pillow>=11.2.1",
    "pydantic>=2.11.7",
    "requests>=2.32.4",
    "smolagents>=1.20.0",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "requests-mock>=1.12.1",
    "ruff>=0.11.4",
]

[project.optional-dependencies]
agentops = [
    "agentops>=0.4.18",
]

[tool.mypy]
strict = true
disallow_untyped_defs = true
exclude = [
    "tests",
]
plugins = []

[tool.ruff.lint]
extend-select = ["I"]
